credit_approval/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
openpyxl==3.1.5
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
